Use HTTPS for Git Clone
I tried the installer inside a fresh VM. I couldn't figure out why the installer didn't install anything until I looked at the script . It was using SSH for the git clone. I didn't have my SSH keys in the VM so it failed. This patch switches the git clone to HTTPS so the clone won't fail for those working in a clean environment.